+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Change Log
 
+## [0.0.33] - 2017-20-07
+### Changed
+
+- CommonJS style requires inside the requireJS blocks are now supported.
+
+    define(function(require) {
+        var moduleB = require('moduleB');
+        moduleB.prop;
+    });
+
 ## [0.0.32] - 2017-06-07
 ### Changed